move sucuri subscriber to new architecture
